indexserver: inline CmdArgs into archiveIndex
We will be adding in a way to index with zoekt-git-index which will
generate arguments in a different way. As such we inline CmdArgs and
change the structure of the tests to test the actual arguments we run
with.
